% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo.calculating_epigenomic_signal.R
\name{flamingo.calculating_epigenomic_signal}
\alias{flamingo.calculating_epigenomic_signal}
\title{flamingo.calculating_epigenomic_signal}
\usage{
flamingo.calculating_epigenomic_signal(epigenomic_file)
}
\arguments{
\item{epigenomic_file}{The path to the epigenomic file, which should has four columns (chr,start,end,score).}
}
\value{
Write out a list of files containing the genomic location and averaged epigenomic signals for DNA fragments.
}
\description{
Calculate the averaged epigenomic signal for each DNA fragment.
}
\keyword{FLAMINGO}
