% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo.construct_flamingo_from_mcool.R
\name{flamingo.construct_flamingo_from_mcool}
\alias{flamingo.construct_flamingo_from_mcool}
\title{flamingo.construct_flamingo_from_mcool}
\usage{
flamingo.construct_flamingo_from_mcool(
  mcool_file,
  normalization,
  resolution,
  chr_name,
  alpha = -0.25,
  n_row = 45000
)
}
\arguments{
\item{mcool_file}{Path to the .mcool file.}

\item{normalization}{Method of normalization.}

\item{resolution}{Target resolution.}

\item{chr_name}{chromosome name}

\item{alpha}{convertion factor between interaction frequency and distance, default is -0.25}

\item{n_row}{number of rows to use the large matrix mode, default to 45000}
}
\value{
A FLAMINGO object containing the resulted Interaction Frequency (IF) matrix, Pairwise Distance (PD) matrix and number of fragments.
}
\description{
Generate FLAMINGO input data from the .mcool format data.
}
\keyword{FLAMINGO}
