% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/flamingo.construct_flamingo_from_sparse_matrix.R
\name{flamingo.construct_flamingo_from_sparse_matrix}
\alias{flamingo.construct_flamingo_from_sparse_matrix}
\title{flamingo.construct_flamingo_from_sparse_matrix}
\usage{
flamingo.construct_flamingo_from_sparse_matrix(
  raw_count_file,
  norm_file = NULL,
  resolution,
  chr_name,
  chr_size,
  alpha = -0.25
)
}
\arguments{
\item{raw_count_file}{Path to the RawCount file from Rao et al.}

\item{norm_file}{Path to the noralization vector file from Rao et al.}

\item{resolution}{Target resolution.}

\item{chr_name}{Name of the chromosome, e.g. chr1.}

\item{chr_size}{Chromosome size}

\item{alpha}{convertion factor between interaction frequency and distance, default is -0.25}
}
\value{
A FLAMINGO object containing the resulted Interaction Frequency (IF) matrix, Pairwise Distance (PD) matrix and number of fragments.
}
\description{
Generate FLAMINGO input data from the sparse format of Hi-C data from Rao et al 2014.
}
\keyword{FLAMINGO}
