% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo.divide_domain.R,
%   R/flamingo.estimate_alpha.R
\name{flamingo.divide_domain}
\alias{flamingo.divide_domain}
\title{flamingo.divide_domain}
\usage{
flamingo.divide_domain(flamingo_obj)

flamingo.divide_domain(flamingo_obj)
}
\arguments{
\item{flamingo_obj}{FLAMINGO object containing the 5kb resolution interaction frequency matrix and pairwise distance matrix.}

\item{domain_res}{Size of the large domain.}

\item{frag_res}{Size of the small DNA fragment.}
}
\value{
Write out a list of Interaction Frequency matrix (IF) and Pairwise Distance matrix (PD)

An alpha value.
}
\description{
Divide the whole chromosome into large domains to perform hierarchical reconstruction

Estimate an alpha value based on the input Hi-C data.
}
\examples{
flamingo.divide_domain(flamingo_obj=flamingo_data,domain_res=1e6,frag_res=5e3)
flamingo.estimate_alpha(flamingo_obj=flamingo_data)
}
\keyword{FLAMINGO}
