% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo.estimate_alpha.R
\name{flamingo.estimate_alpha}
\alias{flamingo.estimate_alpha}
\title{flamingo.estimate_alpha}
\usage{
flamingo.estimate_alpha(flamingo_obj)
}
\arguments{
\item{flamingo_obj}{FLAMINGO object containing the 5kb resolution interaction frequency matrix and pairwise distance matrix.}
}
\value{
An alpha value.
}
\description{
Estimate an alpha value based on the input Hi-C data.
}
\keyword{FLAMINGO}
