% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo.main_func_large.R
\name{flamingo.main_func_large}
\alias{flamingo.main_func_large}
\title{flamingo.main_func_large}
\usage{
flamingo.main_func_large(
  hic_data_low,
  file_format,
  domain_res,
  frag_res,
  chr_size,
  chr_name,
  normalization,
  downsampling_rates,
  lambda,
  max_dist,
  nThread,
  alpha = -0.25,
  max_iter = 500,
  hic_data_high = NULL,
  norm_low = NULL,
  norm_high = NULL,
  n_row = 45000
)
}
\arguments{
\item{hic_data_low}{Input Hi-C data in .hic format, .mcool format or sparse matrix format.}

\item{file_format}{foramt of the input hic data. Could be .hic, .mcool or sparse matrix format.}

\item{domain_res}{Size of the domains in bps, e.g. 1e6.}

\item{frag_res}{Size of the fragment in bps, e.g. 5e3.}

\item{chr_size}{Size of the chromosome in bps.}

\item{chr_name}{Name of the chromosome, e.g. chr1.}

\item{normalization}{Normalization method.}

\item{downsampling_rates}{Fraction of contacts to be used during the reconstruction.}

\item{lambda}{Lagrigian coefficient.}

\item{max_dist}{Maximum allowed distance betwee two consecutive points.}

\item{nThread}{Number of thread avalable for the reconstruction.}

\item{alpha}{Convertion factor between interaction frequency and pairwise distance. default -0.25.}

\item{max_iter}{Maximum iteration for the assembling algorithm. default 500.}

\item{hic_data_high}{Optional. The high resolution HiC data in sparse matrix format. Only required if the file_format is 'sparse matrix'.}

\item{norm_low}{Optional. The normalization vector for the low resolution Hi-C data. Only required if the file_format is 'sparse matrix'.}

\item{norm_high}{Optional. The normalization vector for the high resolution Hi-C data. Only required if the file_format is 'sparse matrix'.}

\item{n_row}{Optional. Number of rows to use the large matrix format. reduce the number if memory is limited. default t0 45000 rows.}
}
\value{
A data.frame containing the FLAMINGO predicted 3D structure.
}
\description{
Main function of FLAMINGO, a wraper for all steps for super large data, i.e. 1kb Hi-C
}
\keyword{FLAMINGO}
