% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo.reconstruct_backbone_structure.R
\name{flamingo.reconstruct_backbone_structure}
\alias{flamingo.reconstruct_backbone_structure}
\title{flamingo.reconstruct_backbone_structure}
\usage{
flamingo.reconstruct_backbone_structure(
  flamingo_data_obj,
  sw,
  lambda,
  max_dist,
  nThread = 28
)
}
\arguments{
\item{flamingo_data_obj}{A flamingo object containing the IF and PD matrix of the backbone, e.g 1MB resolution.}

\item{sw}{Downsample rates, suggesting the fraction of the observed data to be used in the model.}

\item{lambda}{The penalty term for the diagnal entries.}

\item{max_dist}{The maximum allowed distance for the diagnal entries.}

\item{nThread}{Number of thread to be used in the model, default is 28.}
}
\value{
A flamingo_prediction object containing the 3D structure of the backbone.
}
\description{
A wraper of the low-rank matrix completion algorithm to reconstruct the backbone structure
}
\keyword{FLAMINGO}
