% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo.reconstruct_structure_wraper.R
\name{flamingo.reconstruct_structure_wraper}
\alias{flamingo.reconstruct_structure_wraper}
\title{flamingo.reconstruct_structure_wraper}
\usage{
flamingo.reconstruct_structure_wraper(
  index,
  sw,
  lambda,
  max_dist,
  nThread = 28
)
}
\arguments{
\item{index}{Index of a domain}

\item{sw}{Downsample rates, suggesting the fraction of the observed data to be used in the model.}

\item{lambda}{The penalty term for the diagnal entries.}

\item{max_dist}{The maximum allowed distance for the diagnal entries.}

\item{nThread}{Number of thread to be used in the model, default is 28.}
}
\value{
A list of flamingo_prediction object
}
\description{
A wraper of the low-rank matrix completion algorithm to process all the domains
}
\keyword{FLAMINGO}
