% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iflamingo.reconstruct_structure_worker.R
\name{iflamingo.reconstruct_structure_worker}
\alias{iflamingo.reconstruct_structure_worker}
\title{iflamingo.reconstruct_structure_worker}
\usage{
iflamingo.reconstruct_structure_worker(
  input_if,
  pd,
  dist_epi,
  sw,
  lambda,
  lambda_epi,
  max_dist,
  nThread = 2
)
}
\arguments{
\item{input_if}{Input interaction frequency matrix.}

\item{pd}{Input pairwise distance matrix.}

\item{dist_epi}{Input inputed distance based on the epigenomic features}

\item{sw}{Downsample rates, suggesting the fraction of the observed data to be used in the model.}

\item{lambda}{The penalty term for the diagnal entries.}

\item{lambda_epi}{The penalty term for the inputed distances.}

\item{max_dist}{The maximum allowed distance for the diagnal entries.}

\item{nThread}{Number of thread to be used in the model, default is 28.}
}
\value{
A flamingo_prediction object containing the fragment id and 3D coordinates
}
\description{
Reconstruct the 3D genome structure using iFLAMINGO algorithm
}
\keyword{FLAMINGO}
