% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.vtk.R
\name{write.vtk}
\alias{write.vtk}
\title{write.vtk}
\usage{
write.vtk(points, lookup_table, name, opt_path)
}
\arguments{
\item{points}{3D coordiantes predicted by FLAMINGO in x,y,z format.}

\item{lookup_table}{The annotation for each point, could be labels or scores, i.e. TAD annotations.}

\item{name}{output file name}

\item{opt_path}{output file path}
}
\value{
Write out a .vtk file for visualization using Paraview
}
\description{
Convert the FLAMINGO predicted 3D genome structure into .vtk format for visualization
}
\keyword{FLAMINGO}
