% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo.divide_domain.R
\name{flamingo.divide_domain}
\alias{flamingo.divide_domain}
\title{flamingo.divide_domain}
\usage{
flamingo.divide_domain(flamingo_obj, domain_res, frag_res)
}
\arguments{
\item{flamingo_obj}{FLAMINGO object containing the chromosome-wide interaction frequency matrix and pairwise distance matrix.}

\item{domain_res}{Size of the large domain.}

\item{frag_res}{Size of the small DNA fragment.}
}
\value{
Write out a list of Interaction Frequency matrix (IF) and Pairwise Distance matrix (PD)
}
\description{
Divide the whole chromosome into large domains to perform hierarchical reconstruction
}
\keyword{FLAMINGO}
